+++ modules/stream_manager/src/include/FileRead.hpp
@@ -34,7 +34,7 @@
 //=============================================================================
 namespace Nelson {
 NLSSTREAM_MANAGER_IMPEXP ArrayOf
-FileRead(Evaluator* eval, File* fp, int64 sizeToRead, Class classPrecision, size_t skip,
+FileRead(File* fp, int64 sizeToRead, Class classPrecision, size_t skip,
     bool bIsLittleEndian, int& sizeReallyRead);
 };
 //=============================================================================